+++ Source/VirtualRowing/Private/ToolchainBluetoothProbe.h
@@ -1,0 +1,7 @@
+#pragma once
+
+class FToolchainBluetoothProbe final
+{
+  public:
+	static void Run();
+};